pyproject.toml
[tool.poetry]
name = "microservice-summary-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
nameko = "^2.14.1"
transformers = "^4.45.1"
datasets = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
